refactor(br41a): extract portable auth (StorageAdapter + injected fetch) into cowork-bridge; chrome consumes it
